algo-trading/monitor/src/alert_manager.py
import yaml

class AlertManager:
    def __init__(self):
        with open('config/monitoring_config.yml', 'r') as f:
            self.config = yaml.safe_load(f)
        self.alerts = self.config['alerts']

    def check_alerts(self, metrics):
        triggered_alerts = []
        for alert in self.alerts:
            if self.is_alert_triggered(alert, metrics):
                triggered_alerts.append(alert)
        return triggered_alerts

    def is_alert_triggered(self, alert, metrics):
        # Here you would implement the logic to check if an alert is triggered
        # based on the current metrics and the alert configuration
        # This is a placeholder implementation
        return False

    def send_alert(self, alert):
        # Here you would implement the logic to send an alert
        # (e.g., send an email, SMS, or post to a chat system)
        print(f"Alert triggered: {alert['name']}")