algo-trading/monitor/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pyyaml==5.4.1
prometheus_client==0.11.0
pymongo==3.12.0
pandas==1.3.3
matplotlib==3.4.3